Refactor kpf_utils; write out KPF activities.
Change kpf_utils to separate creation of optional packets from writing
them; tweak kpf_geometry to follow.

This is necessary because when we read in an activity record, any kv
packets automatically become activity attributes (rather than peer
packets.) So when writing, we need to find all the optional packets
on the row, assign the kv packets to the activity, and write out
the others as peers.

Also add code to write out KPF activities.